style(InputCarNames): seperate getCarName methods from Application Class and make new class
